Enabled pedantic warning level in project.yml. Added KineticConnectionConfig which is integrated/inherited into KineticConnection to simplify public API and internal handling. HMAC calculation/validation is broken and system tests are currently failing as a result.

+++ src/lib/kinetic_logger.c
@@ -57,7 +57,7 @@ void KineticLogger_Log(const char* message)
     }
 
     fd = LogToConsole ? stderr : fopen(LogFile, "a");
-    if (fd >= 0)
+    if ((int)fd >= 0)
     {
         fprintf(fd, "%s\n", message);
         fflush(fd);
